datos

+++ database/migrations/2024_01_17_141225_create_alumnos_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('alumnos', function (Blueprint $table) {
-            $table->increments('id');
+            $table->unsignedBigInteger('id');
             $table->string('matricula');
             $table->string('nombre', 100);
             $table->string('apellido', 100);
